Several improvements to the subtitle dowloader dialog
* It now fetches the subtitle list from the server on a new worker
thread. No more locking up the player when opening the dialog!
* The dialog now uses the native interfaces available to NT5 or later
to sort the columns.
* The dialog is now 500px wide by default.
* The default columns' widths were changed to something more sensible.
They now properly occupy the whole ListView width without showing
scrollbars.
* The columns' widths will revert to the default size if the registry
setting is corrupted. Previously it always took the player down.
* Reduced flicker.
* Tooltips will now be shown for clipped text.
* Added a StatusBar to notify the user how many subtitles were found,
if it is downloading the list, parsing the list, and if the connection
was unsuccessful.

git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4734 10f7b99b-c216-0410-bff0-8a66a9350fd8