FLACSource: Fix a crash with some corrupted file.
Ensure the buffer is big enough before reading the audio frame. If the buffer is too small, the file is probably corrupted but we try to decode it anyway. There can be some audio glitches but the file is likely to play quite correctly in the end.

This commit fixes ticket #2685.