build.bat: Fix bug introduced by 520944bf5fdf
Commit 520944bf5fdf accidently reversed the if/else-check for InnoSetup and 7z.
This caused release packaging to break on both 32bit and 64bit windows.

This fix works around this by introducing a new variable %os_type% which stores
Win32 or Win64, and uses this to determine the path of InnoSetup and 7z.

The variable %x64_type%, which decided on whether to use the amd64 native or
the x86_amd64 32bit toolchain for a 64 bit build was removed completely,
because visual studio always uses x64_amd64. Hence hardcode this value when
setting ARCH=.