Fix: Post-resize shaders weren't working with VMR-9 (renderless).
The required interface wasn't put in the interfaces list and was still available by chance for the EVR renderers. As a consequence, the m_pMFVDC interface wasn't available for some EVR renderers while to should have been. To keep the behavior consistent with older builds, the order of some checks for m_pMFVDC has been changed.

Also remove the hack that was used for MadVR since it's now correctly fixed.

This commit fixes ticket #2839.