"Play > Filters" menu: simplify the items' state handling.
This menu is dynamically built so there is no reason to use an extra function to set the state of its items. The states are now set correctly when building the menu.

Note that thanks to this change the new UI style is now used for the radio bullet.

This commit fixes ticket #1620.