Merge https://github.com/Isopod/mpc-hc
* Add support for MKV cropping tag

The Matroska container features a tag for "soft" cropping of video
streams (e.g. to compensate for black bars in the video stream without
reencoding). VLC already supports this.

I implemented this by adjusting the frame size used for calculations
in CMainFrame::MoveVideoWindow. A desirable side effect of this method
is that the user can "undo" the cropping by simply zooming out via Pan&Scan.