Internal LAVFilters: Give more flexibility for the path from where to load the filters.
If a "LAVFilters" folder exists together with the executable, it will always be preferred. However if the filters aren't available in that folder but are registered on the computer, they will be loaded from the registered path if that option was enabled on compile time.

This option is disabled by default.