Use VS2012 as the default for the builds.
We've been using VS2012 for the nightlies for some time now without any real issues apart from the renderer one which was reported and worked around.
This will allow us to use more C++11 features.

No files are renamed or removed at this point; we'll do this after 1.7.0 is out.