Fix the DVD favorites: the position in the DVD was not restored.
After 1c3d80162511fe5ddbe36a7988e610cc2ee85a2c the DVD state was restored when the crash was still stopped so it had no effect.

Additionally if "Remember DVD position" is enabled and a DVD favorite has a position, the position from the favorite will be used instead of the remembered position.

This commit fixes #3988.