Audio Switcher: Support fallback to another media type.
If we had accepted an input mediatype that can't be handled by the downstream filter but another compatible mediatype was also proposed by the input pin, we renegotiate the connection with the upstream filter.

This enables audio decoders to fallback to normal decoding if bitstreaming isn't supported.

This commit fixes #3437.