Subresync bar: Make the list control virtual for faster loading.
The list control is just a view for the subtitle data so there is no need to duplicate it into the list control.

This makes loading a subtitle into the Subresync bar nearly instant even for complex subtitle files. Such files will now be order of magnitude faster to load.