Add option to set jpg quality when saving images.
Added to the save dialog used for screenshots and thumbnails. Default is
set to 90, which is a change from what it used to be (which was what the
Windows encoder defaulted to, 75).

On Vista or later, it's only shown if JPEG encoding is selected, and
hidden if not. On XP, it's simply grayed out if something other than JPEG
is selected.

Closes #138.