Add "Play next file in the folder" event to single time events menu.
From my understanding people doesn't care about loops and even
persistence of the after playback switches, they just want to have it
easy accessible and executed on playlist end. This isn't fully consistent
with every time events which respect loop counter, but it doesn't have to
be, we ought to look at it as two separate features which provide
different possible workflows. This way majority of the users should be happy,
both this who wants to use looping feature and those who wants to abuse
"play next" feature.

Fixes #4971