Pause the video during the MediaInfo analysis if playing from an optical drive.
Optical drives are so bad at handling concurrent accesses that playback would be disrupted anyway.

Instead, we do the analysis synchronously when the user chooses to display the MediaInfo tab. Because the analysis is deferred until it is really needed, it is possible that the filter graph is destroyed before doing the analysis and that another media is playing. This will not cause any issue but playback will be paused anyway to avoid adding an extra check to know whether we are currently playing from the same optical drive.

Fixes #5127.